Massive overhaul of featurisation and visualisation.
Atom labels are no longer based on proximity to a ligand atom of
interest, but on their participation or otherwise in actual bonding
interactions, as determined by the PLIP library. As such, many of the
new functions are modified versions of those found in that repository.

Visualisation now also integrates plip, such that the pymol objects
produced include all bonding interactions as well as colouring of
receptor atoms based on probability labels.